requirements.txt
streamlit >= 1.30.0
pandas >= 2.0.0
google-genai >= 1.0.0
beautifulsoup4 >= 4.12.0
requests >= 2.31.0
google-api-python-client >= 2.100.0
google-auth-oauthlib >= 1.1.0
google-auth-httplib2 >= 0.1.0
python-dateutil >= 2.8.2
lxml >= 4.9.0
groq >= 0.4.2
